Remove transitive govmomi inclusion from guest.UUID
